FE API: move priceByPaymentQuery to the frontend-api package
