+++ packages/frontend-api/src/Resources/config/graphql-types/PaymentDecorator.types.yaml
@@ -22,7 +22,11 @@ PaymentDecorator:
             price:
                 type: "Price!"
                 description: "Payment price"
-                resolve: '@=query("priceByPaymentQuery", value)'
+                resolve: '@=query("priceByPaymentQuery", value, args["cartUuid"], context)'
+                args:
+                    cartUuid:
+                        type: "Uuid"
+                        defaultValue: null
             images:
                 type: "[Image!]!"
                 description: "Payment images"
