+++ project-base/app/config/graphql/types/ModelType/Payment/Payment.types.yaml
@@ -7,14 +7,6 @@ Payment:
             type:
                 type: "String!"
                 description: "Type of payment"
-            price:
-                type: "Price!"
-                description: "Payment price"
-                resolve: '@=query("priceByPaymentQuery", value, args["cartUuid"], context)'
-                args:
-                    cartUuid:
-                        type: "Uuid"
-                        defaultValue: null
             images:
                 type: "[Image!]!"
                 description: "Payment images"
