+++ project-base/app/src/FrontendApi/Resolver/Price/PriceQuery.php
@@ -4,66 +4,24 @@ declare(strict_types=1);
 
 namespace App\FrontendApi\Resolver\Price;
 
-use App\FrontendApi\Model\Cart\CartFacade;
 use App\FrontendApi\Resolver\Price\Exception\ProductPriceMissingUserError;
-use App\Model\Order\Preview\OrderPreviewFactory;
 use App\Model\Product\Product;
 use ArrayObject;
-use Shopsys\FrameworkBundle\Component\Domain\Domain;
-use Shopsys\FrameworkBundle\Model\Customer\User\CurrentCustomerUser;
-use Shopsys\FrameworkBundle\Model\Payment\Payment;
-use Shopsys\FrameworkBundle\Model\Payment\PaymentPriceCalculation;
-use Shopsys\FrameworkBundle\Model\Pricing\Currency\CurrencyFacade;
 use Shopsys\FrameworkBundle\Model\Pricing\Price;
 use Shopsys\FrameworkBundle\Model\Product\Pricing\ProductPrice;
-use Shopsys\FrameworkBundle\Model\Product\ProductCachedAttributesFacade;
-use Shopsys\FrameworkBundle\Model\Product\ProductOnCurrentDomainFacadeInterface;
 use Shopsys\FrameworkBundle\Model\Transport\Transport;
-use Shopsys\FrameworkBundle\Model\Transport\TransportPriceCalculation;
 use Shopsys\FrontendApiBundle\Component\GqlContext\GqlContextHelper;
-use Shopsys\FrontendApiBundle\Model\Price\PriceFacade;
 use Shopsys\FrontendApiBundle\Model\Resolver\Price\PriceQuery as BasePriceQuery;
 
 /**
  * @property \App\Model\Product\ProductOnCurrentDomainElasticFacade $productOnCurrentDomainFacade
+ * @property \App\Model\Customer\User\CurrentCustomerUser $currentCustomerUser
+ * @property \App\Model\Order\Preview\OrderPreviewFactory $orderPreviewFactory
+ * @method __construct(\Shopsys\FrameworkBundle\Model\Product\ProductCachedAttributesFacade $productCachedAttributesFacade, \App\Model\Product\ProductOnCurrentDomainElasticFacade $productOnCurrentDomainFacade, \Shopsys\FrameworkBundle\Model\Payment\PaymentPriceCalculation $paymentPriceCalculation, \Shopsys\FrameworkBundle\Component\Domain\Domain $domain, \Shopsys\FrameworkBundle\Model\Pricing\Currency\CurrencyFacade $currencyFacade, \Shopsys\FrameworkBundle\Model\Transport\TransportPriceCalculation $transportPriceCalculation, \Shopsys\FrontendApiBundle\Model\Price\PriceFacade $priceFacade, \App\Model\Customer\User\CurrentCustomerUser $currentCustomerUser, \App\Model\Order\Preview\OrderPreviewFactory $orderPreviewFactory, \Shopsys\FrontendApiBundle\Model\Cart\CartApiFacade $cartApiFacade)
+ * @method \Shopsys\FrameworkBundle\Model\Pricing\Price calculateIndependentPaymentPrice(\App\Model\Payment\Payment $payment)
  */
 class PriceQuery extends BasePriceQuery
 {
-    /**
-     * @param \Shopsys\FrameworkBundle\Model\Product\ProductCachedAttributesFacade $productCachedAttributesFacade
-     * @param \App\Model\Product\ProductOnCurrentDomainElasticFacade $productOnCurrentDomainFacade
-     * @param \Shopsys\FrameworkBundle\Model\Payment\PaymentPriceCalculation $paymentPriceCalculation
-     * @param \Shopsys\FrameworkBundle\Component\Domain\Domain $domain
-     * @param \Shopsys\FrameworkBundle\Model\Pricing\Currency\CurrencyFacade $currencyFacade
-     * @param \Shopsys\FrameworkBundle\Model\Transport\TransportPriceCalculation $transportPriceCalculation
-     * @param \Shopsys\FrontendApiBundle\Model\Price\PriceFacade $priceFacade
-     * @param \App\FrontendApi\Model\Cart\CartFacade $cartFacade
-     * @param \App\Model\Customer\User\CurrentCustomerUser $currentCustomerUser
-     * @param \App\Model\Order\Preview\OrderPreviewFactory $orderPreviewFactory
-     */
-    public function __construct(
-        ProductCachedAttributesFacade $productCachedAttributesFacade,
-        ProductOnCurrentDomainFacadeInterface $productOnCurrentDomainFacade,
-        PaymentPriceCalculation $paymentPriceCalculation,
-        Domain $domain,
-        CurrencyFacade $currencyFacade,
-        TransportPriceCalculation $transportPriceCalculation,
-        PriceFacade $priceFacade,
-        private readonly CartFacade $cartFacade,
-        private readonly CurrentCustomerUser $currentCustomerUser,
-        private readonly OrderPreviewFactory $orderPreviewFactory,
-    ) {
-        parent::__construct(
-            $productCachedAttributesFacade,
-            $productOnCurrentDomainFacade,
-            $paymentPriceCalculation,
-            $domain,
-            $currencyFacade,
-            $transportPriceCalculation,
-            $priceFacade,
-        );
-    }
-
     /**
      * @param \App\Model\Transport\Transport $transport
      * @param string|null $cartUuid
@@ -84,7 +42,7 @@ class PriceQuery extends BasePriceQuery
             return parent::priceByTransportQuery($transport);
         }
 
-        $cart = $this->cartFacade->findCart($customerUser, $cartUuid);
+        $cart = $this->cartApiFacade->findCart($customerUser, $cartUuid);
 
         if ($cart === null) {
             return parent::priceByTransportQuery($transport);
@@ -103,51 +61,6 @@ class PriceQuery extends BasePriceQuery
 
         return $this->transportPriceCalculation->calculatePrice(
             $transport,
-            $currency,
-            $orderPreview->getProductsPrice(),
-            $domainId,
-        );
-    }
-
-    /**
-     * @param \App\Model\Payment\Payment $payment
-     * @param string|null $cartUuid
-     * @param ?\ArrayObject $context
-     * @return \Shopsys\FrameworkBundle\Model\Pricing\Price
-     */
-    public function priceByPaymentQuery(
-        Payment $payment,
-        ?string $cartUuid = null,
-        ?ArrayObject $context = null,
-    ): Price {
-        $cartUuid = $cartUuid ?? GqlContextHelper::getCartUuid($context);
-
-        /** @var \App\Model\Customer\User\CustomerUser|null $customerUser */
-        $customerUser = $this->currentCustomerUser->findCurrentCustomerUser();
-
-        if ($customerUser === null && $cartUuid === null) {
-            return parent::priceByPaymentQuery($payment);
-        }
-
-        $cart = $this->cartFacade->findCart($customerUser, $cartUuid);
-
-        if ($cart === null) {
-            return parent::priceByPaymentQuery($payment);
-        }
-
-        $domainId = $this->domain->getId();
-        $currency = $this->currencyFacade->getDomainDefaultCurrencyByDomainId($domainId);
-        $orderPreview = $this->orderPreviewFactory->create(
-            $currency,
-            $domainId,
-            $cart->getQuantifiedProducts(),
-            null,
-            $payment,
-            $customerUser,
-        );
-
-        return $this->paymentPriceCalculation->calculatePrice(
-            $payment,
             $currency,
             $orderPreview->getProductsPrice(),
             $domainId,
